Refactor: mejor uso de los tipados que ya existen y acomodo de nombres
