+++ src/app/inicio/comercializador/administracionUsuarios/formulario/types/formulario.tsx
@@ -1,145 +1,19 @@
-export interface RolesInterface {
-  id: string;
-  nombre: string;
-  nombreNormalizado: string
-}
+import RolesInterface from "@/app/inicio/usuarios/interfaces/RolesInterface";
+import RefEmpleador from "@/app/inicio/usuarios/interfaces/RefEmpleador";
+import CargoInterface from "@/app/inicio/usuarios/interfaces/CargoInterface";
+import type {
+  UsuarioFormFields as BaseUsuarioFormFields,
+  RequestMethod as UsuarioRequestMethod,
+  ValidationErrors as BaseValidationErrors,
+  TouchedFields as BaseTouchedFields,
+  Props as UsuarioFormProps,
+} from "@/app/inicio/usuarios/UsuarioForm";
 
-export interface RefEmpleador {
-  interno: number;
-  nombre1: string;
-  nombre2: string;
-  razonSocial: string;
-      // Add other fields if needed
-}
+export type RequestMethod = UsuarioRequestMethod;
+export type UsuarioFormFields = BaseUsuarioFormFields;
 
-export interface CargoInterface {
-  id: number;
-  descripcion: string; 
-}
-
-export type RequestMethod = "create" | "edit" | "view" | "delete" | "activate" | "remove";
-
-export interface UsuarioFormFields {
-  nombre: string;
-  email: string;
-  cuit: string; // Keep as string for form input, will convert to number on submit
-  phoneNumber: string;
-  matricula?: string;
-  fechaNacimiento?: string;
-  canalInterviniente?: string;
-  inicioFecha?: string;
-  bajaFecha?: string | null;
-  domicilioCalle?: string;
-  domicilioNro?: string;
-  domicilioPiso?: string;
-  domicilioEntreCalle1?: string;
-  domicilioEntreCalle2?: string;
-  domicilioCodPostal?: string;
-  domicilioCodLocalidad?: string;
-  domicilioLocalidad?: string;
-  domicilioProvincia?: string;
-  cargoId?: number;
-  password?: string;
-  confirmPassword?: string;
-  rol: string;
-  // tipo: string;
-  userName: string;
-  empresaId: number;
-  id?: string;
-  comision?: number;
-  serviciosAdicionales?: number;
-  aplicaIva?: number;
-  srtComercializadorOrganizadorInterno?: number;
-}
-
-export interface Props {
-  open: boolean;
-  onClose: () => void;
-  onSubmit: (formData: UsuarioFormFields) => void;
-  roles: RolesInterface[];
-  cargos?: CargoInterface[];
-  refEmpleadores?: RefEmpleador[];
-  initialData?: UsuarioFormFields;
-  errorMsg?: string | null;
-  method: RequestMethod; // MODO DE OPERACIÓN
-  isSubmitting?: boolean;
-  isAdmin?: boolean; // Nuevo parámetro para determinar si el usuario es admin
-}
+export type Props = UsuarioFormProps;
 
 // Interfaces completas para errores y campos tocados
-export interface ValidationErrors {
-  nombre?: string;
-  email?: string;
-  cuit?: string;
-  phoneNumber?: string;
-  matricula?: string;
-  fechaNacimiento?: string;
-  domicilioCalle?: string;
-  domicilioNro?: string;
-  domicilioPiso?: string;
-  domicilioEntreCalle1?: string;
-  domicilioEntreCalle2?: string;
-  domicilioCodPostal?: string;
-  domicilioCodLocalidad?: string;
-  domicilioLocalidad?: string;
-  domicilioProvincia?: string;
-  cargoId?: string;
-  rol?: string;
-  // tipo?: string;
-  // userName?: string;
-  id?: string;
-}
-
-export interface TouchedFields {
-  nombre?: boolean;
-  email?: boolean;
-  cuit?: boolean;
-  phoneNumber?: boolean;
-  matricula?: boolean;
-  fechaNacimiento?: boolean;
-  canalInterviniente?: boolean;
-  inicioFecha?: boolean;
-  bajaFecha?: boolean;
-  domicilioCalle?: boolean;
-  domicilioNro?: boolean;
-  domicilioPiso?: boolean;
-  domicilioEntreCalle1?: boolean;
-  domicilioEntreCalle2?: boolean;
-  domicilioCodPostal?: boolean;
-  domicilioCodLocalidad?: boolean;
-  domicilioLocalidad?: boolean;
-  domicilioProvincia?: boolean;
-  cargoId?: boolean;
-  rol?: boolean;
-  comision?: boolean;
-  serviciosAdicionales?: boolean;
-  aplicaIva?: boolean;
-  // tipo?: boolean;
-  // userName?: boolean;
-  id?: boolean;
-}
-
-export type ComercializadorGetAll = {
-  interno?: number;
-  Interno?: number;
-  cuil?: number;
-  CUIL?: number;
-  matricula?: string;
-  email?: string;
-  telefono?: string;
-  movil?: string;
-  referenteDatosInterno?: number;
-  canalInterviniente?: string;
-  inicioFecha?: string;
-  bajaFecha?: string;
-  comision?: number;
-  aplicaIva?: number;
-  serviciosAdicionales?: number;
-  srtComercializadorOrganizadorInterno?: number;
-  [key: string]: unknown;
-}
-
-export type ComercializadorApiResponse = {
-  data: ComercializadorGetAll[];
-
-};
+export type ValidationErrors = BaseValidationErrors;
+export type TouchedFields = BaseTouchedFields;
